# api/openapi.yaml
openapi: 3.1.0
info:
  title: Qdrant Web Backup API
  description: |
    REST API for managing Qdrant vector database backups.
    Provides endpoints for cluster monitoring, collection management,
    snapshot operations, storage configuration, and job tracking.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.qdrant-backup.example.com
    description: Production server
...
paths:
...
  /api/v1/collections/{name}/snapshots:
    get:
      tags:
        - Snapshots
      summary: List collection snapshots
      description: Returns all snapshots for a specific collection
      operationId: listSnapshots
      parameters:
        - name: name
          in: path
          required: true
          description: Collection name
          schema:
            type: string
      responses:
        '200':
          description: Snapshots list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsListResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Snapshots
      summary: Create collection snapshot
      description: Creates a new snapshot for the specified collection
      operationId: createSnapshot
      parameters:
        - name: name
          in: path
          required: true
          description: Collection name
          schema:
            type: string
        - name: wait
          in: query
          required: false
          description: Wait for snapshot creation to complete
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Snapshot created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Snapshot creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/collections/{name}/snapshots/{snapshotName}:
    get:
      tags:
        - Snapshots
      summary: Download snapshot
      description: Downloads a snapshot file
      operationId: downloadSnapshot
      parameters:
        - name: name
          in: path
          required: true
          description: Collection name
          schema:
            type: string
        - name: snapshotName
          in: path
          required: true
          description: Snapshot name
          schema:
            type: string
      responses:
        '200':
          description: Snapshot file stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Snapshot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Snapshots
      summary: Delete snapshot
      description: Deletes a snapshot from the collection
      operationId: deleteSnapshot
      parameters:
        - name: name
          in: path
          required: true
          description: Collection name
          schema:
            type: string
        - name: snapshotName
          in: path
          required: true
          description: Snapshot name
          schema:
            type: string
        - name: wait
          in: query
          required: false
          description: Wait for deletion to complete
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Snapshot deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: Snapshot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/collections/{name}/snapshots/recover:
    post:
      tags:
        - Snapshots
      summary: Recover collection from snapshot
      description: Restores a collection from a snapshot URL or local file
      operationId: recoverSnapshot
      parameters:
        - name: name
          in: path
          required: true
          description: Collection name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoverSnapshotRequest'
      responses:
        '200':
          description: Recovery started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoverResponse'
        '400':
          description: Invalid recovery parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/storage/config:
    get:
      tags:
        - Storage
      summary: Get storage configurations
      description: Returns all configured storage backends
      operationId: getStorageConfig
      responses:
        '200':
          description: Storage configurations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageConfigListResponse'

    post:
      tags:
        - Storage
      summary: Create storage configuration
      description: Creates a new storage backend configuration
      operationId: createStorageConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorageConfigCreate'
      responses:
        '200':
          description: Storage configuration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageConfigResponse'
        '400':
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/storage/config/{id}:
    put:
      tags:
        - Storage
      summary: Update storage configuration
      description: Updates an existing storage backend configuration
      operationId: updateStorageConfig
      parameters:
        - name: id
          in: path
          required: true
          description: Storage configuration ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorageConfigUpdate'
      responses:
        '200':
          description: Storage configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageConfigResponse'
        '404':
          description: Configuration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Storage
      summary: Delete storage configuration
      description: Deletes a storage backend configuration
      operationId: deleteStorageConfig
      parameters:
        - name: id
          in: path
          required: true
          description: Storage configuration ID
          schema:
            type: string
      responses:
        '200':
          description: Storage configuration deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '400':
          description: Cannot delete default configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Configuration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/storage/test:
    post:
      tags:
        - Storage
      summary: Test storage connectivity
      description: Tests connectivity to a storage backend without saving
      operationId: testStorageConnectivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorageConfigCreate'
      responses:
        '200':
          description: Connectivity test completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageTestResponse'
...
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok]
        time:
          type: number
          format: float
          description: Request processing time in seconds
      required:
        - status
        - time
    ErrorResponse:
      type: object
      properties:
        status:
          type: object
          properties:
            error:
              type: string
              description: Error message
          required:
            - error
        time:
          type: number
          format: float
      required:
        - status
        - time
    DeleteResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            result:
              type: boolean
              description: True if deletion was successful

    # Cluster schemas
...
    Snapshot:
      type: object
      properties:
        name:
          type: string
          description: Snapshot filename
        creation_time:
          type: string
          format: date-time
          description: Snapshot creation timestamp
        size:
          type: integer
          description: Snapshot size in bytes
        checksum:
          type: string
          description: Snapshot checksum
      required:
        - name
        - creation_time
        - size
    SnapshotsListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            result:
              type: array
              items:
                $ref: '#/components/schemas/Snapshot'
    SnapshotResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            result:
              $ref: '#/components/schemas/Snapshot'
    RecoverSnapshotRequest:
      type: object
      properties:
        location:
          type: string
          description: Snapshot URL (S3, HTTP, or file://)
        priority:
          type: string
          enum: [replica, snapshot, no_sync]
          default: snapshot
          description: Recovery priority for conflict resolution
        api_key:
          type: string
          description: Optional API key for authenticated URLs
      required:
        - location
    RecoverResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            result:
              type: boolean
              description: True if recovery started successfully

    # Storage schemas
    StorageType:
      type: string
      enum: [local, s3]
      description: Storage backend type
    StorageConfig:
      type: object
      properties:
        id:
          type: string
          description: Configuration ID
        name:
          type: string
          description: Configuration name
        type:
          $ref: '#/components/schemas/StorageType'
        path:
          type: string
          description: Local filesystem path (for local type)
        s3_endpoint:
          type: string
          format: uri
          description: S3 endpoint URL (for s3 type)
        s3_bucket:
          type: string
          description: S3 bucket name (for s3 type)
        s3_region:
          type: string
          description: S3 region (for s3 type)
        s3_access_key:
          type: string
          description: S3 access key (masked in responses)
        is_default:
          type: boolean
          description: Whether this is the default storage
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - name
        - type
        - is_default
        - created_at
        - updated_at
    StorageConfigCreate:
      type: object
      properties:
        name:
          type: string
          description: Configuration name
        type:
          $ref: '#/components/schemas/StorageType'
        path:
          type: string
          description: Local filesystem path (required for local type)
        s3_endpoint:
          type: string
          format: uri
          description: S3 endpoint URL (required for s3 type)
        s3_bucket:
          type: string
          description: S3 bucket name (required for s3 type)
        s3_region:
          type: string
          description: S3 region (required for s3 type)
        s3_access_key:
          type: string
          description: S3 access key (required for s3 type)
        s3_secret_key:
          type: string
          description: S3 secret key (required for s3 type)
        is_default:
          type: boolean
          default: false
          description: Set as default storage
      required:
        - name
        - type
    StorageConfigUpdate:
      type: object
      properties:
        name:
          type: string
          description: Configuration name
        path:
          type: string
          description: Local filesystem path
        s3_endpoint:
          type: string
          format: uri
          description: S3 endpoint URL
        s3_bucket:
          type: string
          description: S3 bucket name
        s3_region:
          type: string
          description: S3 region
        s3_access_key:
          type: string
          description: S3 access key
        s3_secret_key:
          type: string
          description: S3 secret key
        is_default:
          type: boolean
          description: Set as default storage
    StorageConfigListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            result:
              type: array
              items:
                $ref: '#/components/schemas/StorageConfig'
    StorageConfigResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            result:
              $ref: '#/components/schemas/StorageConfig'
    StorageTestResult:
      type: object
      properties:
        success:
          type: boolean
          description: Whether connectivity test passed
        message:
          type: string
          description: Test result message
      required:
        - success
    StorageTestResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            result:
              $ref: '#/components/schemas/StorageTestResult'

    # Job schemas
...